Feature: New Redis detector and storage option (#5)
A Redis solution for detection and storage is now available to allow Anicetus
to be used on distributed systems. The Redis connection will rely on the
redigo client library [1].

The library API was also modified to allow the context to be injected to cancel
the action during I/O operations. This required a module bump.

The Go compiler was also upgraded to 1.24.

The token bucket algorithm was modified to apply a penalty once no more tokens
are available. The penalty means that the bucket will be drained anytime some
request is made without available tokens, requiring the caller to wait a full
token generation period to send another request. To satisfy this behaviour we
had to copy and modify golang.org/x/time/rate@v0.10.0.

[1] https://github.com/gomodule/redigo